requirements.txt
deepgram-sdk>=3.7,<4
google-genai>=0.3
python-dotenv>=1.0
httpx>=0.27
